requirements.txt
catboost==1.2.7
joblib==1.3.2
matplotlib==3.8.2
numpy==2.3.2
optuna==4.4.0
pandas==2.3.2
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.4.1.post1
scipy==1.16.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
xgboost==3.0.4
